+++ CHANGELOG.md
@@ -3,6 +3,14 @@
 
 ### BREAKING CHANGES
 
+- Renamed `corsOptions.methods` to `corsOptions.allowedMethods`.
+- Renamed `corsOptions.credentials` to `corsOptions.allowCredentials`.
+
+<a name="1.2.0"></a>
+## [1.2.0](https://github.com/ts-stack/ditsmod/releases/tag/1.2.0) (2022-10-04)
+
+### BREAKING CHANGES
+
 - Restored `preflightContinue` options.
 
 <a name="1.1.0"></a>